Delegated the event processor to a seperate function for future threading support.
Fixed a bug in which Multi samples would load into memory as strings rather than bytes, resulting in an invalid null character in the sample data.
Changed the user interface to use a dual-channel volume visualizer and a panning gauge.
Updated the channel header to include the type of output.