Fixed several interface bugs.
engine.py
Renamed `Event` to `Command` for brevity.

interface.py
Fixed a bug in which channel was 1-wider than user-specified.
Fixed a bug in which a panning of 127 would crash the interface.
Fixed a bug in which the tempo label would go unused.
Fixed a bug in which the notes display would overflow the column.

parser.py
Added 0xCA command conditional in the loop address getter.