Fix a player bug with null notes.
Fixed a player bug in which notes would flood a track's note queue if
the track had a NULL voice and would attempt to play after the track's
first call to `VOICE`. The note queue is now cleared every cycle of
`Player.play_notes` when a track has a NULL voice.

Additionally replaced the `queue.Queue` use for the track note queue
with a `deque` for ease of use (no multi-threading in this program yet,
so why bother?)

Additionally replaced all tuple conversions of a track's `track_data`
dictionary keys and values with instance variables declared at track
initialization.